Get_abs_path complete
